refactor: update kotlin Upload DSL
